api/requirements.txt
fastapi
pydantic
openai
uvicorn
requests
aioredis
tortoise-orm
aiomysql
python-multipart